# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BaseForge API",
    "description": "AI-Ready Intelligence Layer for the Base Ecosystem. Real-time DeFi analytics, protocol risk scoring, whale tracking, and compressed context payloads for AI agents.",
    "version": "1.0.0",
    "contact": { "name": "BaseForge", "url": "https://github.com/AmnAnon/baseforge-v1" },
    "license": { "name": "MIT", "url": "https://github.com/AmnAnon/baseforge-v1/blob/main/LICENSE" }
  },
  "servers": [
    { "url": "https://baseforge-v1.vercel.app", "description": "Production (Vercel)" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": { "type": "apiKey", "in": "header", "name": "X-API-Key", "description": "API key for agent endpoints" },
      "AdminAuth": { "type": "apiKey", "in": "header", "name": "x-admin-key", "description": "Admin key for management endpoints" }
    },
...
  },
  "paths": {
...
    "/api/portfolio": {
      "get": {
        "summary": "Wallet portfolio balances",
        "tags": ["Portfolio"],
        "parameters": [{ "name": "address", "in": "query", "required": true, "schema": { "type": "string" }, "description": "0x wallet address" }],
        "responses": { "200": { "description": "Portfolio positions" }, "400": { "description": "Invalid address" } }
      }
    },
...
  }
}